[GPU] Refactoring and test fixes
