+++ src/plugins/intel_gpu/src/graph/network.cpp
@@ -325,8 +325,9 @@ network::network(engine& engine,
 network::network(engine& engine,
                  const std::set<std::shared_ptr<program_node>>& nodes,
                  const ExecutionConfig& config,
+                 std::shared_ptr<InferenceEngine::CPUStreamsExecutor> task_executor,
                  bool is_internal)
-    : network(program::build_program(engine, nodes, config, is_internal), config, engine.create_stream(config), is_internal) {}
+    : network(program::build_program(engine, nodes, config, task_executor, is_internal), config, engine.create_stream(config), is_internal) {}
 
 network::network(program::ptr program, uint16_t stream_id)
     : network(program, program->get_config(), program->get_engine().create_stream(program->get_config()), false, stream_id == 0) {}
@@ -531,8 +532,9 @@ network::ptr network::build_network(engine& engine,
 network::ptr network::build_network(engine& engine,
                                     const std::set<std::shared_ptr<program_node>>& nodes,
                                     const ExecutionConfig& config,
+                                    std::shared_ptr<InferenceEngine::CPUStreamsExecutor> task_executor,
                                     bool is_internal) {
-    return std::make_shared<network>(engine, nodes, config, is_internal);
+    return std::make_shared<network>(engine, nodes, config, task_executor, is_internal);
 }
 
 void network::validate_primitives() {
