+++ src/plugins/intel_gpu/src/plugin/graph.cpp
@@ -45,38 +45,39 @@ using namespace InferenceEngine::details;
 namespace ov {
 namespace intel_gpu {
 
-Graph::Graph(InferenceEngine::CNNNetwork& network, RemoteContextImpl::Ptr context, ExecutionConfig exec_config, uint16_t stream_id)
+Graph::Graph(InferenceEngine::CNNNetwork& network, RemoteContextImpl::Ptr context, const ExecutionConfig& config, uint16_t stream_id)
     : m_context(context)
     , m_networkName(network.getName())
-    , m_exec_config(exec_config)
+    , m_config(config)
     , m_stream_id(stream_id)
     , m_state(0) {
-    m_program = std::make_shared<Program>(network, get_engine(), exec_config);
+    m_program = std::make_shared<Program>(network, get_engine(), config);
     if (m_program->m_max_batch > 1)
-        m_exec_config.set_property(ov::intel_gpu::max_dynamic_batch(m_program->m_max_batch));
+        m_config.set_property(ov::intel_gpu::max_dynamic_batch(m_program->m_max_batch));
     Build();
 }
 
-Graph::Graph(cldnn::BinaryInputBuffer &ib, RemoteContextImpl::Ptr context, ExecutionConfig exec_config, uint16_t stream_id)
+Graph::Graph(cldnn::BinaryInputBuffer &ib, RemoteContextImpl::Ptr context, const ExecutionConfig& config, uint16_t stream_id)
     : m_context(context)
-    , m_exec_config(exec_config)
+    , m_config(config)
     , m_stream_id(stream_id)
     , m_state(0) {
-    m_program = std::make_shared<Program>(get_engine(), exec_config);
+    m_program = std::make_shared<Program>(get_engine(), config);
     if (m_program->m_max_batch > 1)
-        m_exec_config.set_property(ov::intel_gpu::max_dynamic_batch(m_program->m_max_batch));
+        m_config.set_property(ov::intel_gpu::max_dynamic_batch(m_program->m_max_batch));
 
     ib >> m_program->inputLayouts;
     ib >> primitiveIDs;
     ib >> outputDims;
 
-    m_networks.emplace_back(std::make_shared<cldnn::network>(ib, get_engine().create_stream(exec_config), get_engine(), m_stream_id));
+    m_networks.emplace_back(std::make_shared<cldnn::network>(ib, get_engine().create_stream(config), get_engine(), m_stream_id));
 }
 
 Graph::Graph(std::shared_ptr<Graph> graph, uint16_t stream_id)
         : m_context(graph->m_context)
         , m_program(graph->m_program)
         , m_networkName(graph->m_networkName)
+        , m_config(graph->m_config)
         , m_stream_id(stream_id)
         , m_state(0) {
     Build();
@@ -138,10 +139,10 @@ std::shared_ptr<cldnn::network> Graph::BuildNetwork(std::shared_ptr<cldnn::progr
 
     auto externalQueue = m_context->get_external_queue();
     if (externalQueue) {
-        if (m_exec_config.get_property(ov::num_streams) != 1)
+        if (m_config.get_property(ov::num_streams) != 1)
             IE_THROW(ParameterMismatch) << "Throughput streams can't be used with shared queue!\n";
         auto &engine = m_program->get_engine();
-        network = std::make_shared<cldnn::network>(program, engine.create_stream(m_exec_config, externalQueue), m_stream_id);
+        network = std::make_shared<cldnn::network>(program, engine.create_stream(m_config, externalQueue), m_stream_id);
     } else {
         network = std::make_shared<cldnn::network>(program, m_stream_id);
     }
@@ -170,7 +171,7 @@ Graph::variable_states_map Graph::AllocateVariablesMemories() {
 std::shared_ptr<ngraph::Function> Graph::GetExecGraphInfoByPrimitivesInfo(std::vector<cldnn::primitive_info>& primitives_info,
                                                                           bool filter_const_primitives) {
     OV_ITT_SCOPED_TASK(itt::domains::intel_gpu_plugin, "Graph::GetExecGraphInfoByPrimitivesInfo");
-    if (m_exec_config.get_property(ov::enable_profiling)) {
+    if (m_config.get_property(ov::enable_profiling)) {
         try {
             // Update may throw an exception for step-by-step runtime graph dump,
             // since network->get_executed_primitives() method can't be called before network execution
