+++ src/plugins/intel_gpu/src/plugin/ops/parameter.cpp
@@ -197,7 +197,7 @@ static void CreateParameterOp(Program& p, const std::shared_ptr<ngraph::op::v0::
         p.inputLayouts.insert({ inputInfo->name(), networkInputLayout });
         p.add_primitive(*op, cldnn::input_layout(inputName, networkInputLayout));
     } else {
-        if (ColorFormat::NV12 == preProcess.getColorFormat() && p.get_exec_config().get_property(ov::intel_gpu::nv12_two_inputs)) {
+        if (ColorFormat::NV12 == preProcess.getColorFormat() && p.get_config().get_property(ov::intel_gpu::nv12_two_inputs)) {
             // for NV12, create two input layouts with reorder instead of one,
             // and then would expect compound blob in inferRequest
             if (InferenceEngine::Layout::NCHW != l &&
